Rename module from logic to utils.
We use utils in more places
